Cleanup properties file handling and localize ide messages
