+++ kt/plugins/godot-intellij-plugin/src/main/resources/META-INF/plugin.xml
@@ -22,7 +22,6 @@
         <moduleBuilder builderClass="godot.intellij.plugin.module.GodotModuleBuilder"/>
         <notificationGroup id="Error message" displayType="STICKY_BALLOON" key="notification.group.error" isLogByDefault="true"/>
         <annotator implementationClass="godot.intellij.plugin.annotator.clazz.RegisterClassAnnotator" language="JVM"/>
-        <annotator implementationClass="godot.intellij.plugin.annotator.clazz.TestAnnotator" language="JVM"/>
         <annotator implementationClass="godot.intellij.plugin.annotator.function.RegisterFunctionAnnotator" language="JVM"/>
         <annotator implementationClass="godot.intellij.plugin.annotator.function.RpcAnnotator" language="JVM"/>
         <annotator implementationClass="godot.intellij.plugin.annotator.property.RegisterPropertiesAnnotator" language="JVM"/>
@@ -37,7 +36,7 @@
             extensions="tscn"
             implementationClass="godot.intellij.plugin.language.scene.filetype.GodotSceneFileType"/>
         <editorNotificationProvider
-            implementation="godot.intellij.plugin.notification.FileIndexingNotification" />
+            implementation="godot.intellij.plugin.notification.SettingsFetchingNotification" />
         <externalSystemTaskNotificationListener
                 implementation="godot.intellij.plugin.gradle.GradleSystemListener"/>
     </extensions>
