+++ kt/plugins/godot-intellij-plugin/src/main/resources/messages/generalLabels.properties
@@ -83,3 +83,14 @@ codeVision.settings.enableCheckbox=[GodotKotlinJVM] Display registered name abov
 codeVision.registeredName.tooltip=Click to copy registered name
 codeVision.registeredName.text=Registered as {0}
 codeVision.registeredName.notRegisteredBecauseIsAbstract.text=Will not be registered, as it's abstract. Members are registered through inheriting classes
+settingsIndex.action.title=Importing {0} GodotKotlinJvm
+settingsIndex.error.connectionFailed=Connecting with the GodotKotlinJvm plugin failed: Code checks may be incorrect.
+settingsIndex.error.compatibility.semVerFailure=Plugin 'GodotKotlinJvm' is incompatible with the current version of the GodotKotlinJvm Gradle plugin. Upgrade the version of the gradle plugin com.utopia-rise.godot-kotlin-jvm
+settingsIndex.error.compatibility.idePluginOutOfDate=Plugin 'GodotKotlinJvm' is no longer compatible with the current version of the GodotKotlinJvm Gradle plugin. Use version {0} or later.
+settingsIndex.error.compatibility.gradlePluginOutOfDate=Gradle plugin 'com.utopia-rise.godot-kotlin-jvm' is incompatible with the current GodotKotlinJvm IntelliJ Plugin version. Use version {0} or later.
+settingsIndex.unconfigured.reason.gradleSyncInProgress=GodotKotlinJvm is waiting for Gradle to finish syncing.
+settingsIndex.unconfigured.reason.syncInProgress=GodotKotlinJvm is setting up...
+settingsIndex.error.showErrorAction=Show error
+settingsIndex.error.connectionError=Error connecting to GodotKotlinJvm.
+settingsIndex.error.copyAction=Copy
+settingsIndex.error.copyMessage=Stacktrace copied to clipboard!
